\name{gwm}
\alias{gwm}
\docType{data}
\title{ Internal Data Used by map.world Function }
\description{
  It stores longitudes and latitudes used for drawing the world map.
}
\usage{data(gwm)}
\format{
  A data frame with 6920 observations on the following 2 variables.
  \describe{
    \item{\code{lon}}{longitudes on earth}
    \item{\code{lat}}{latitudes on earth}
  }
}
\source{
  S Archive under \url{http://lib.stat.cmu.edu}
}
\examples{
data(gwm)
}
\keyword{datasets}
