\name{rmp}
\alias{rmp}
\docType{data}
\title{ Internal Data Used by map.world Function }
\description{
  It stores vertex index info used for drawing the world map.
}
\usage{data(rmp)}
\format{
  A data frame with 54 observations on the following variable.
  \describe{
    \item{\code{inc}}{a numeric vector}
  }
}
\source{
  S Archive under \url{http://lib.stat.cmu.edu}
}
\examples{
data(rmp)
}
\keyword{datasets}
